pyproject.toml
[project]
name = "privacy-assessment-agent"
version = "0.1.0"
description = "An AI agent to assess privacy policies against regulations."
authors = [
    {name = "User", email = "user@example.com"}, # Placeholder
]
dependencies = [
    "google-adk",
    "google-generativeai",
    "requests",
    "beautifulsoup4",
    "python-dotenv",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.urls]
Homepage = "https://example.com/privacy_agent" # Placeholder

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["privacy_agent*"] # Will find the 'privacy_agent' package directory
